HSEARCH-3589 Introduce a common superclass for object fields and value fields in the Lucene metamodel
We'll need this in the next commits, so as to build a list of children
to expose in the metamodel.
